This should be the final re-do of point-vs-vector typing. Details below.
This is a thorough cleanup of this library's headers, comments, and data types. The revised, and final, philosophy is that JavaFX Point2D and Point3D are only used when doing pure geometry, containment tests, mouse click positions, etc. Once in 3D land (this doesn't include the linear objects), everything should be expressed as vectors as most of the work involves vector math, offsets, additions, etc.